bot/requirements.txt
psycopg2-binary
aiogram